+++ docs/RELEASE-NOTES.md
@@ -1,5 +1,17 @@
 # Release Notes
 
+## 2.3.0 - ?? July 2020
+
+Breaking changes:
+- (None)
+
+New features:
+- Added DictionaryExtensions.ContainsAllKey
+- Added DictionaryExtensions.ContainsAnyKey
+
+Bug fixes / internal changes:
+- (None)
+
 ## 2.2.0 - 13 July 2020
 
 Breaking changes: